ai_pr_reviewer_final/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==1.12.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0
gitpython==3.1.31